+++ pyhamtools/callsign_exceptions.py
@@ -1,4 +1,4 @@
-
+"""In this files, strange special callsigns are located which can't be resolved with any other rule"""
 
 callsign_exceptions={
     '7QAA' : '7Q',